i2c-gpio: add sda_enforce_dir option
Currently, i2c-gpio assumes that it can read the SDA gpio even when
configured as an output pin, which is not true for all GPIO controllers.

This patch adds an option to i2c-gpio to enable the use of the driver
with GPIO controllers that require pins to be configured as outputs
before writing to them and as inputs before reading from them.
Such controllers need to set the sda_enforce_dir option.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>